(PC-14097) feat(firebase): move maintenance types to types.ts
