+++ src/libs/firestore/types.ts
@@ -10,3 +10,27 @@ export enum RemoteStoreDocuments {
   MAINTENANCE_MESSAGE = 'message',
   MINIMAL_BUILD_NUMBER = 'minimalBuildNumber',
 }
+
+export enum MAINTENANCE {
+  UNKNOWN = 'UNKNOWN',
+  OFF = 'OFF',
+  ON = 'ON',
+}
+
+export type Maintenance =
+  | {
+      status: MAINTENANCE.UNKNOWN
+      message: undefined
+    }
+  | {
+      status: MAINTENANCE.OFF
+      message: undefined
+    }
+  | {
+      status: MAINTENANCE.ON
+      message: string
+    }
+
+export type OnMaintenanceChange = (maintenance: Maintenance) => void
+
+export type Unsubscribe = () => void
